external-events: Rearchitect into a module.
