+++ external-events/src/mod.rs
@@ -1,0 +1,8 @@
+//! This crate manages all external event reporting, including 1) price feeds from centralized
+//! exchanges, 2) StarkWare events, including nullifier reveals in order to hang up MPCs, and 3)
+//! Ethereum events, like sequencer rotation or L1 deposits.
+mod errors;
+mod exchange_handlers;
+pub mod exchanges;
+pub mod reporter;
+pub mod tokens;
